requirements.txt
dj-database-url==0.5.0
Django==3.2.7
psycopg2-binary==2.9.1
gunicorn==20.1.0 
whitenoise==5.3.0
cloudinary==1.26.0
djangorestframework==3.12.2
django-cors-headers==3.8.0 
django-cloudinary-storage==0.3.0 
Pillow==8.3.1
django-filter==2.4.0
djangorestframework-simplejwt==4.7.2
